requirements.txt
pysrt==0.21.0
google-generativeai==0.1.0